+++ src/test/kotlin/com/atlassian/performance/tools/virtualusers/LoadTestTest.kt
@@ -8,6 +8,7 @@ import com.atlassian.performance.tools.virtualusers.api.config.VirtualUserBehavi
 import com.atlassian.performance.tools.virtualusers.api.config.VirtualUserTarget
 import com.atlassian.performance.tools.virtualusers.mock.RemoteWebDriverMock
 import com.atlassian.performance.tools.virtualusers.mock.WebElementMock
+import net.jcip.annotations.ThreadSafe
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import org.openqa.selenium.By
@@ -15,22 +16,36 @@ import java.net.URI
 import java.time.Duration
 import java.util.concurrent.atomic.AtomicInteger
 
+@ThreadSafe
 class LoadTestTest {
 
+    /**
+     * Prevents concurrent shared global state mutations if the test methods are concurrent.
+     */
     private val globalStateLock = Object()
 
     @Test
     fun shouldRunLoadTestWithoutExceptions() {
-        val loadTest = loadTest(1)
+        val loadTest = loadTest(
+            virtualUsers = 6,
+            skipSetup = false
+        )
 
         synchronized(globalStateLock) {
+            TestBrowser.reset()
             loadTest.run()
         }
+
+        assertThat(TestBrowser.timesStarted.get()).isEqualTo(7)
+        assertThat(TracingScenario.setup).isEqualTo(true)
     }
 
     @Test
     fun shouldInstallOnlyOnce() {
-        val loadTest = loadTest(20)
+        val loadTest = loadTest(
+            virtualUsers = 20,
+            skipSetup = false
+        )
 
         synchronized(globalStateLock) {
             MockWebdriverRuntime.reset()
@@ -40,8 +55,26 @@ class LoadTestTest {
         assertThat(MockWebdriverRuntime.installations.get()).isEqualTo(1)
     }
 
+    @Test
+    fun shouldSkipSetup() {
+        val loadTest = loadTest(
+            virtualUsers = 4,
+            skipSetup = true
+        )
+
+        synchronized(globalStateLock) {
+            TestBrowser.reset()
+            TracingScenario.reset()
+            loadTest.run()
+        }
+
+        assertThat(TestBrowser.timesStarted.get()).isEqualTo(4)
+        assertThat(TracingScenario.setup).isEqualTo(false)
+    }
+
     private fun loadTest(
-        virtualUsers: Int
+        virtualUsers: Int,
+        skipSetup: Boolean
     ): LoadTest = LoadTest(
         options = VirtualUserOptions(
             target = VirtualUserTarget(
@@ -49,7 +82,7 @@ class LoadTestTest {
                 userName = "username",
                 password = "password"
             ),
-            behavior = VirtualUserBehavior.Builder(NoopScenario::class.java)
+            behavior = VirtualUserBehavior.Builder(TracingScenario::class.java)
                 .browser(TestBrowser::class.java)
                 .load(VirtualUserLoad(
                     virtualUsers = virtualUsers,
@@ -57,6 +90,7 @@ class LoadTestTest {
                     ramp = Duration.ZERO,
                     flat = Duration.ofSeconds(1)
                 ))
+                .skipSetup(skipSetup)
                 .build()
         )
     )
@@ -69,9 +103,18 @@ class LoadTestTest {
 
         override fun start(): CloseableRemoteWebDriver {
             driverRuntime.ensureRunning()
+            timesStarted.incrementAndGet()
             return CloseableRemoteWebDriver(
                 driver = TestWebDriver()
             )
+        }
+
+        companion object Counter {
+            val timesStarted: AtomicInteger = AtomicInteger(0)
+
+            fun reset() {
+                timesStarted.set(0)
+            }
         }
     }