+++ src/test/kotlin/com/atlassian/performance/tools/virtualusers/TracingScenario.kt
@@ -7,7 +7,8 @@ import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 import com.atlassian.performance.tools.jiraactions.api.memories.UserMemory
 import com.atlassian.performance.tools.jiraactions.api.scenario.Scenario
 
-internal class NoopScenario : Scenario {
+internal class TracingScenario : Scenario {
+
     private val noopAction = object : Action {
         override fun run() {
         }
@@ -18,10 +19,23 @@ internal class NoopScenario : Scenario {
     }
 
     override fun getSetupAction(jira: WebJira, meter: ActionMeter): Action {
-        return noopAction
+        return object : Action {
+            override fun run() {
+                setup = true
+            }
+        }
     }
 
     override fun getLogInAction(jira: WebJira, meter: ActionMeter, userMemory: UserMemory): Action {
         return noopAction
+    }
+
+    companion object Trace {
+        var setup = false
+            private set
+
+        fun reset() {
+            setup = false
+        }
     }
 }